+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added colors to the help screen
 - Merge all the `--words-red`, `--words-green`, etc. flags into a single `--highlight [COLOR]:[WORDS]` flag
 - Renamed `--no-builtin-keywords` to `--disable-builtin-keywords`
+- Added a --pager flag to override the default pager in addition to the env variable
 
 ## 5.3.0